+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 - Fix serialization of JSONB parameters in prepared statements.
 
-- Preliminary support for the PostgreSQL POINT, LINE, LSEG, BOX and PATH types (geometric types).
+- Preliminary serialization and deserialization support for the PostgreSQL `POINT`, `LINE`, `LSEG`,
+  `BOX`, `PATH` and `POLYGON` types (geometric types).
 
 
 ## [0.36] - 2024-06-23